Change place list to map in LocationNotificationService
