+++ src/main/java/org/tndata/android/compass/activity/PlaygroundActivity.java
@@ -6,7 +6,7 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
 import org.tndata.android.compass.R;
-import org.tndata.android.compass.service.LocationProviderService;
+import org.tndata.android.compass.service.LocationNotificationService;
 import org.tndata.android.compass.ui.button.TransitionButton;
 
 
@@ -30,7 +30,7 @@ public class PlaygroundActivity extends AppCompatActivity implements View.OnClic
 
         state = 0;
 
-        startService(new Intent(this, LocationProviderService.class));
+        startService(new Intent(this, LocationNotificationService.class));
     }
 
     @Override
@@ -55,7 +55,7 @@ public class PlaygroundActivity extends AppCompatActivity implements View.OnClic
 
     @Override
     protected void onDestroy(){
-        LocationProviderService.cancel();
+        LocationNotificationService.cancel();
         super.onDestroy();
     }
 }
